Add URL-Session object.
Add a URL session object to the view controller to prepare for network
connections.  Add delegate method to handle session-level errors.  Add
new error type for the controller’s own errors, with a conversion to
NSError.

Add a conditional call to create an error situation, to test the
delegate method.