Add URL-Session-Task object.
Add an optional URL session-task object to the view controller for the
network connection.  Update the action methods to actually do
connections and disconnections.  Add another delegate method to handle
task-level errors.

Add a constant for the default echo port.  Update how the “connected”
status is determined.  Add a property to generate the server location
data from the text fields; it’s also used to store the location in the
task object.