Add the actual transactions of the Echo protocol.
Add a constant for how many consecutive received empty data blocks are
acceptable.  Expand the list of protocol-related errors.

Add a helper function to display an alert for errors.

Add a method to send out data from the echo text entry field to the
server.  Add a method to receive and compare data from the server to
the text entry field.  Update the Echo action method to call the
echo-send method (which in turn calls the echo-receive method).